Updating GSIGEO2011 to include all the original data
The initial commit was cropping the grid to the extent of EPSG:6695

New the grid provides better results when compared to reference data.
The result of the validation as been included in the README. The grid
has also been flipped vertically as the upper bound was South to the lower
bound.