Modification of no_kv_NKGETRF14_EPSG7922_2000.tif (#79)
The Norwegian Mapping Authority has a modified version of the correction grid no_kv_NKGETRF14_EPSG7922_2000 used in the national ETRS89NO<>ITRF2014 transformation. The grid differs from the first release within 10-15 mm in the north of the country. There will be no more versions of the grid until ITRF2020 is published. We have chosen to not rename the filename in regards to not do any changes in nkg.sql in Proj.